rewrite ffoption
